[Feature] Improve SFTP read speed.
The max packet size is limited by TCP/IP to 65535 but that's not a
bottleneck because the underlying IO code can help us split the
request into multiple packets. The window size is also not a concern
because it's dynamic and only represents the current quota which is
constantly being consumed and allocated.

This change makes our code use an 1 MiB instead 64 KiB buffer, and
increased the speed of copying file from remote from 1 MiB/s to 3
MiB/s in a local network. Also note that using a 4 MiB buffer didn't
give a significant improvement over 1 MiB.